functions done
